requirements.txt
lark-parser==0.12.0
graphviz==0.20.3
click==8.1.7
pytest==8.0.0
psutil
prometheus_client
flask
flask_cors
werkzeug